Refactor keyboard package
In preparation for changes.
