+++ packages/keyboard/src/layouts/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export { EN_US } from './en-US';
